+++ server/src-lib/Hasura/Server/Query.hs
@@ -89,6 +89,8 @@ data RQLQueryV1
   | RQDeleteEventTrigger !DeleteEventTriggerQuery
   | RQRedeliverEvent     !RedeliverEventQuery
   | RQInvokeEventTrigger !InvokeEventTriggerQuery
+  | RQPauseEventTrigger  !PauseEventTriggerQuery
+  | RQResumeEventTrigger  !ResumeEventTriggerQuery
 
   -- query collections, allow list related
   | RQCreateQueryCollection !CreateCollection
@@ -251,6 +253,8 @@ queryModifiesSchemaCache (RQV1 qi) = case qi of
   RQDeleteEventTrigger _          -> True
   RQRedeliverEvent _              -> False
   RQInvokeEventTrigger _          -> False
+  RQPauseEventTrigger _           -> False
+  RQResumeEventTrigger _          -> False
 
   RQCreateQueryCollection _       -> True
   RQDropQueryCollection _         -> True
@@ -376,6 +380,8 @@ runQueryM rq = withPathK "args" $ case rq of
       RQDeleteEventTrigger q       -> runDeleteEventTriggerQuery q
       RQRedeliverEvent q           -> runRedeliverEvent q
       RQInvokeEventTrigger q       -> runInvokeEventTrigger q
+      RQPauseEventTrigger q        -> runPauseEventTrigger q
+      RQResumeEventTrigger q       -> runResumeEventTrigger q
 
       RQCreateQueryCollection q        -> runCreateCollection q
       RQDropQueryCollection q          -> runDropCollection q
@@ -457,6 +463,8 @@ requiresAdmin = \case
     RQDeleteEventTrigger _          -> True
     RQRedeliverEvent _              -> True
     RQInvokeEventTrigger _          -> True
+    RQPauseEventTrigger _           -> True
+    RQResumeEventTrigger _         -> True
 
     RQCreateQueryCollection _       -> True
     RQDropQueryCollection _         -> True